# app/schemas/bets.py
from pydantic import BaseModel, Field
from typing import Optional, List
from datetime import datetime
from uuid import UUID


class BetIn(BaseModel):
    external_id: Optional[str] = None
    user_ref: Optional[str] = None
    match_id: str
    bookmaker_id: UUID
    selection_id: UUID
    stake: float = Field(gt=0)
    price: float = Field(gt=1.0)
    placed_at: datetime
    idempotency_key: Optional[str] = None

class BetCreateOut(BaseModel):
    created: bool
    bet_id: Optional[UUID] = None

class BetsList(BaseModel):
    items: List[BetIn]


BetIn.model_config = {
    "json_schema_extra": {
        "examples": [{
            "external_id": "ticket-123",
            "user_ref": "user42",
            "match_id": "m1",
            "bookmaker_id": "024c6a47-1a14-4549-935f-31e22e747670",
            "selection_id": "bea8671c-e889-4e3d-91d3-b407bc186408",
            "stake": 100.0,
            "price": 2.10,
            "placed_at": "2025-08-18T12:30:00Z",
            "idempotency_key": "abc-123"
        }]
    }
}